+++ public/res/CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ## Version 2.0.30200421 (dev)
+- Replaced element i with span.
 - Cleaned up MaterialHelper class and Drawer class.
 
 ## Version 2.0.30200420 (dev)